Move to constants, make a public
